add club_stadion_page + new pages mechanic
