+++ lib/pages/clubPages/club_stadium_page.dart
@@ -1,0 +1,47 @@
+import 'package:flutter/material.dart';
+import 'package:pocket_eleven/design/colors.dart';
+
+class ClubStadiumPage extends StatefulWidget {
+  const ClubStadiumPage({super.key});
+
+  @override
+  State<ClubStadiumPage> createState() => _ClubStadiumPageState();
+}
+
+class _ClubStadiumPageState extends State<ClubStadiumPage> {
+  late Image _clubStadiumImage;
+
+  @override
+  void initState() {
+    super.initState();
+    _clubStadiumImage = Image.asset('assets/background/club_stadion.png');
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        backgroundColor: AppColors.hoverColor,
+        title: const Text('Club Stadion'),
+      ),
+      body: Column(
+        children: [
+          Container(
+            height: MediaQuery.of(context).size.height / 3,
+            decoration: BoxDecoration(
+              image: DecorationImage(
+                image: _clubStadiumImage.image,
+                fit: BoxFit.cover,
+              ),
+            ),
+          ),
+          Expanded(
+            child: Container(
+              color: AppColors.primaryColor,
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+}
